Application chutes
